backend/requirements.txt
fastapi
uvicorn
python-multipart
groq
chromadb
sentence-transformers
pymupdf
transformers
torch
easyocr
opencv-python
Pillow
numpy
pytest
httpx